+++ src/libu/dictionary.c
@@ -84,9 +84,9 @@ static void * mem_double(void * ptr, int size)
 
 unsigned dictionary_hash(char * key)
 {
-	int			len ;
+	size_t			len ;
 	unsigned	hash ;
-	int			i ;
+	size_t			i ;
 
 	len = strlen(key);
 	for (hash=0, i=0 ; i<len ; i++) {
